pyproject.toml
[project]
name = "agent-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.4.9.2",
    "autogen-ext[grpc,mcp,openai]>=0.4.9.2",
    "mcp-server-fetch>=2025.1.17",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.1",
]

[project.scripts]
agent-core = "main:main"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]

[tool.uv.workspace]
members = [
    "creator",
]
